Re-arrange dags to make them imortable into pipe_reference
